+++ CHANGES.rst
@@ -5,6 +5,11 @@ This is a record of all past ``respy`` releases and what went into them in rever
 chronological order. We follow `semantic versioning <https://semver.org/>`_ and all
 releases are available on `PyPI <https://pypi.org/project/respy/>`_.
 
+2.0.0 - 2019-
+-------------
+
+- `#177 <https://github.com/OpenSourceEconomics/respy/pull/177>`_ removes all Fortran
+  files and ensures that all tests still run through.
 
 1.2.1 - 2019-05-19
 ------------------
@@ -22,7 +27,6 @@ releases are available on `PyPI <https://pypi.org/project/respy/>`_.
   for theses.
 - `#186 <https://github.com/OpenSourceEconomics/respy/pull/186>`_ replaces statsmodels
   as a dependency with our own implementation.
-
 
 1.2.0 - 2019-04-23
 ------------------